refactor: align backend entities with new schema

+++ autowash-backend/src/main/java/com/autowash/dto/AdminVoucherResponse.java
@@ -1,12 +1,13 @@
 package com.autowash.dto;
 
+import com.autowash.entity.enums.DiscountType;
 import java.time.Instant;
 import java.util.List;
 
 public record AdminVoucherResponse(
         String code,
         String discountType,
-        int discountValue,
+        long discountValue,
         long minAmount,
         Instant expiresAt,
         boolean active,
