+++ autowash-backend/src/main/java/com/autowash/dto/ValidateVoucherResponse.java
@@ -1,12 +1,13 @@
 package com.autowash.dto;
 
+import com.autowash.entity.enums.DiscountType;
 import java.time.Instant;
 
 public record ValidateVoucherResponse(
         String voucherCode,
         boolean isValid,
         String discountType,
-        int discountValue,
+        long discountValue,
         long discountAmount,
         long finalAmount,
         Instant expiresAt
